add ability to register/query artifacts with attributes.
Multiple artifacts have been stored so far in a flat namespace,
which mean that it is not possible to dissambiguate each instance.
It is also not possible to get a single one of them, as multiple
artifacts can only be queried as a whole list.

Added the ability to attach attributes to multiple artifacts
registration and query. This will allow users to selectively
obtain the artifacts in a list.

Test: added unit tests and integration tests
Bug: N/A
Relnote: added ability to register and query artifacs with attributes on the variant API

Change-Id: I0b28e585bb17a41c9b762e1e3bdd61e1b3821d64