GradleRule: AndroidComponents support.
With the new fixture, we want to avoid giving access to
the build file as we want to be able to support kts,
groovy and declarative seamlessly. This posese a problem
for the androidComponent extension. This is an extension
that isn't particularly declarative and therefore we
cannot just use the DslProxy mechanism.

This commit introduces a way to create custom plugins
that support the androidComponent extension. Base
plugin classes will delegate to a callback the ability
to act on the extension. Code inside the rule will
gather the callback classes (and their references) and
collect them in a jar. The generated project will use
that jar as a source of custom plugin.

Current limitations:
- only support app and libs
- only one callback per project
- does not yet handle reconfigure. This is a bit trickier
  but should be doable.

Bug: N/A
Test: Updated PluginTest with a full custom logic.
Change-Id: I6d83a37a66e67ffeee8d723abd987bf5dc1dc580