[fused lib] Add experimentalProperties and dependency validation option
Adds experimental option for the ability to disable dependnecy
validation. This is a safeguard in the event we add a new validation
that may block users from building.

We will likely remove this property once the plugin is more stable or
replace with option to disable specific validation rather than
everything.

Fixes: 378080297
Fixes: 378080572
Test: FusedLibraryClassesVerificationTest
Change-Id: I27d496c927492250f5649927cfc2c11231b38877