fix inconsistency in buildConfigFields declaration
Using the out keyword in DeviceTest/AndroidTest is inconsitent with
how it is defined on the variant. This is not a problem for DeviceTest
as it is still an incubating interface.

Since it is not possible to use this correctly with AndroidTest which
is a stable inteface, we can either fix it (clearly people are not
using it in kotlin) or we can leave unchanged and tell people to
move to the DeviceTest interface when using Kotlin.

So far, I have decided to change it in AndroidTest since it is
currently broken in kotlin) but I expect to discuss this
separately.

Test: added regression tests.
Fixes: 369698074
Change-Id: I467f5a2bd796c169b0485ed81ebfe5b15af1e576