GradleRule: quality of life API changes.
Put a default project path for both app and lib.
This is mostly useful for the app since there is often
only one.

the default androidApplication project creation now
creates an empty manifest to allow building these
projects (there's an opt-out)

Added androidJavaApplication and androidKotlinApplication
in case a full app is needed, but with the empty manifest
created by default, this is less needed (this can be
useful for building bundles that require more like
version code)

Bug: N/A
Test: updated the integ tests
Change-Id: Ida422837bd1f9fce1d43a75c8ba64db612f6aad0