Change default artifact file name from "out.jar" to "${artifactType}.dat"
When the user of the Artifacts API doesn't specify a file name for a
file artifact, we need to provide a default file name.
 - In commit fed6f0b, we chose the default file name to be "out"
 - Then in commit 8b886ca, we changed it to "out.jar"

The issue is that "out.jar" will be incorrect for non-jar files
(e.g., text files).

This commit changes the default file name to be "${artifactType}.dat" as
".dat" is commonly regarded as a generic extension representing any kind
of file contents.

Bug: 151076862
Test: Updated
Change-Id: I7cd2911a32e5bb6b0d79d8c2e5dbff1ba030cf98