Disable dependency resolution check for project isolation
AGP has a check to ensure dependencies are not resolved during
configuration phase.

In Gradle 8.11, when *project isolation* is enabled (default mode for
our tests), we see that many dependencies are resolved during
configuration (https://github.com/gradle/gradle/issues/31483).

To unblock the Gradle update, the previous commit 62f1a52 disabled that
check when *configuration cache* is enabled, but that was too strong.

This commit re-enables the check for *configuration cache* and disables
it for *project isolation*.

Bug: 374810264 (Follow up)
Test: Existing
Change-Id: I1ad81312e9cd843bf08a9cbda09e5ddb1cbb0e45