GradleRule: custom PrivacySandboxSdkDefinition
Using AndroidProjectDefinition did not make sense
as the Privacy sandbox does not have a component
extension.

This chnage also removed the Android prefix for these
classes as it's a bit unnecessary.

Finally, this change adds a BaseAndroidProject layer
so that AndroidProject and other, different, android
projects can share some method (mostly output access)

Bug: N/A
Test: existing tests
Change-Id: Ia6352a4d9993a9febd5d52e2efd9c667a0844cdc