Add apk installation lifecycle hook.
Some user tasks might want to be registered as dependent of the
installation task, to verify the APK for instance. Providing the
ability to register such tasks.

This is the first part of the fix required by the referenced bug
which wants the ability to register tasks that depend on the
install task. This other direction will be provided in a subsequent
CL.

Test: unit test and integration test
Bug: 199194011
Change-Id: I6d834a554dcedb6d89e5101f7abcf70418335153