Introduce DefaultAnnotationValuesAnnotationAttributesProvider
... that retrieves default values from an annotation's class.

In multipreview-asm, we fetch the preview parameters that are
defined by the user. However, if the user doesn't provide
a value, we end up with an empty value instead of the
default one specified in the preview annotation's class.

DefaultAnnotationValuesAnnotationAttributesProvider will
allow us to fallback onto those default values when the user
hasn't overridden any of the annotation's parameters.

Bug: 377900045
Test: added DefaultAnnotationValuesAnnotationAttributesProviderTest
Change-Id: Id4535d12b64f9892cc0e961bae9c6c1d64e8f45d