GradleRule: convert more KGP tests
this commit also adds reset methods to
reset the state of the DSL in the definition
so that reconfigure() start from scratch if
needed. This is useful if there are DSL calls
that are hard to revert or override.

This also fixes an issue with running with
jvmToolchain(17) as the location to find
such VM was missing. This is injected from
the rules as the definition itself cannot
hardcode this support or the unit test will
fail (as the JDK is not available when
running the integ/framework tests). This new
support also rework handling of the custom
plugins during reconfigure instead of caching
it all locally in each project.

Bug: N/A
Test: converting tests to new fixtures
Change-Id: I42405c1a58bf666177549bc1acc13b222035ff86