Update lint issue docs to include release version
This CL updates the lint issue documentation generation tool to include
(on the issue details page) information on which specific version the
individual issue was added in, as well as the release month and year.

It only includes specific canary/beta release information for recent
issues that haven't been included in a stable release yet; for other
issues, it just lists the stable version.

It used to include the "copyright year", found on the detector class
implementation. This is less accurate, especially for detectors that
have many new issues added over the years, where the original issue year
was also listed on new issues.  For example, the WebpUnsupported
detector was previously listed as coming from 2011 (since it's part of
the original IconDetector source file), but is now adjusted to 2017 when
it was really introduced.

Test: Manual inspection of results on
   https://googlesamples.github.io/android-custom-lint-rules/checks/index.md.html
Fixes: 385875832
Change-Id: I3d0c8db40da905e6d0ca8bd95831b51ce3265d0f