Add caches in AnnotationHandler
AnnotationHandler may recalculate the relevant annotations for the same
element repeatedly, so we add caches. When profiling the time spent in
getRelevantAnnotations on one example file, the caches appear to save
about 500ms (for that one file). The file does not appear to have many
annotations, but its AST is quite large, with nested lambdas and
variable references.

Bug: 382396705
Test: existing AnnotationHandlerTest
Change-Id: I0f9f9409d4792844fec61a06533185bcea288e0f