GradleRule: Convert more Privacy Sandbox tests
These privacy sandbox tests validates the result of the
variant ApkOutputProviders API.

The test must use the variant API to register a task,
wire it to the variant's output providers, and use code
in the task to validate the content that is returned by
the API (since that content is used at runtime inside
a task).

The previous tests used a multitude of string content
concatenated together to generate this scipt inside
a build file. Each string handled a different part,
allowing to combine various parts for various scenarios
(app module, dynamic feature module, test component,
etc...).

These new tests replicate this via a hierarchy of
classes and interfaces allowing one to combine
different callbacks and different verification
tasks.

This also improve significantely the verification
code to not simply check boolean results, but
offers a richer output message in case of errors
to allow better debugging.

Bug: N/A
Test: converting tests to new fixtures
Change-Id: Ia9987ebd2921530c0439e3ba001460d0b610d832