Use correct path for androix.inspection.inspection classes
The jar in the m2 repository only contains sources. The classes are
contained in the aar, so they can't be added automatically to the lib in
intellij, which prevents the code depending on this lib form compiling
in the ide.

The classes are extraded from the aar to a folder in prebuilt. This
change updates the intellij library to get the classes from the right
location.

Test: NA
Bug: NA
Change-Id: Ie02eeead3c2f3da0ed950ba6fcc2ddf1fea6092f