[SDK Index] Store deprecated issues even if flag is false
... so tests that change this flag can do so even after initialization.

We have tests that check that the issues are not shown if the flag is
flase and are shown if it is true and those are working as expected in
this branch (the default is true and some tests change it to false) but
for branches that the default is false and tests change it to true there
are issues. The problem is that during initialization, GooglePlaySdkIndex
would not save the information of deprecated library issues if
showDeprecationIssues was false.

This change saves the information of deprecated library issues
regardless of the value of showDeprecationIssues.

Bug: 381276797
Test: SdkIndexIssuesSyncListenerTest
Change-Id: Ib9604319f6de22a67671fc62f2bbb92330361875