Read AndroidVersion directly from target hash
Remove over-complicated mechanism that parses the target hash into an
AndroidVersion, then writes the components of that AndroidVersion as
string keys in properties, and then reassembles AndroidVersion on
demand within AvdInfo.

Instead, just store the target hash in the config.ini and parse it when
needed. After this change, there's no unique information stored in the
metadata ini files.

Bug: 389973543
Test: AvdManagerTest
Change-Id: I5404138adef5c4343511f0d279f7b814cfdad152