Revert "[jacoco] Use @Classpath for JacocoTransform inputs"
This reverts commit 44f580899071edfd8114639b685b9a5d07b69f92.

Reason for revert: This is the cause of b/390736538, as using @Classpath
can result in a modified event being split into two an add event and a
delete, and if the delete happens second the output will be deleted.

We could fix this in a more elaborate way, but the safest thing seems to
be a revert for now.

Bug: 250664038
Fixes: 390736538
Test: Manual for now
Change-Id: Ibc7d51816801ee3e7665c4c75292f72eb28c2931