Make KmpVariant extend LibraryVariant
To make transition less messy, KmpVariantImpl will keep implementing
the previous (existing) internal classes, instead of following the
structured implementation pattern in which case KmpVariantImpl would
need to implement VariantImpl. This is because VariantImpl class has
various inputs that are not applicable to kmp. Iteratively we will try
to migrate as much as possible to VariantImpl so that we avoid
implementing new features twice in the future.

Bug: 380218409
Test: existing
Change-Id: I5889ca9f148614d4a48711371afcce08586a1ac3