refactor WithLazyInitialization
The annotation WithLazyInitialization was tagged to be put on
constructors and providing the target method through an
annotation parameter. This made it a bit brittle and not type
safe as the method name was expressed as a String which is
unsafe.

Instead, tag the annotation as a Function annotation and at
the DSL decoration time, generate a method call for each method
annotated. The usefulness of supporting multiple methods can be
argued but I see no value in restricting it either.

Test: Existing
Bug: N/A
Change-Id: Iaa467cd39e4fb5223882dc86b66e91720a6920e6