MultiZip support for AarSubject
Made Zip an abstract task in order to be able to
override getEntry/getEntries. This allows the
creation of views over one or more Zip instance.

First usage is a view over multiple zip files
to get access to all the classes and java resources
in one go (from main jar and secondary jars)

This change also cleans up the JarSubject to
be more consistent as a container of binary
class names.

Finally, this change also fix an issue where the
zip coming from the GradleRule fixture was
not properly closed.

Bug: N/A
Test: updated some tests + new unit tests
Change-Id: I57f3450a8f2b2c71bdaa9c012b200081f4d7e9e4