JarSubject is not a ZipSubject anymore
Since JarSubject handles resournces and classes
in its own way, also having the ZipSubject API
(like entries(), textFile, innerZip) can be confusing.

Therefore JarSubject does not extend AbstractZipSubject
anymore. However, since we wanted to keep exits/doesNotExist,
we moved these 2 to a newer parent class.

This change highlighted another potential trap when using
AarSubject (which is a ZipSubject): inside an lambda acting
on a JarSubject, the ZipSubject API would be available, on
the AarSubject parent scope. This could be confusing.
Therefore, we introduce a DslMarker to avoid these kinds of
risks.

finally, this change also removes the secondaryJar(s) API
as it's really not necessary, and not used anywhere.

Bug: N/A
Test: updated tests
Change-Id: Id30e9811f342bbd76f4cf5a3964970aec1906e50