Zip/Subject improvements
renamed FilteredZip and MultiZip to make it clearer
that they are a view over a zip. Also moved innerZip()
to simpleZip only as there's no need for these views
to have their own caching since underlying zip
does it for them. This allows us to make these
views not be auto-closeable so the API is more
clear about what should be closed.

Minor other tweaks and fixes, including making
the API more Java friendly

Bug: N/A
Test: existing tests
Change-Id: I8b1914647d11209c0de9554e37cfee09bbd94584