Allow call to constructor with value class parameter
Technically, a primary constructor whose value parameter has a value
class type becomes `private` in order to prevent it from being accessed
by Java. If the call-site is Kotlin, a synthetic constructor with a
marker is created/used, but that is not visible to Java either, i.e.,
absence of `public` synthetic constructor in LC modeling.
Therefore, UAST/PSI resolution ended up with the original private
constructor, which causes false positives (for usages in Kotlin).

This commit adds an extra bail-out logic that checks
1) if the call-site is Kotlin; and
2) if the resolved call is a constructor with `value` class parameter.

Bug: 396584142
Test: unit test added
Change-Id: I7e903ca4c70c990935f547d3a6c815fd91c3ca4d