Rearrange GoogleMavenArtifactId
Make it an implementation of a new interface,
WellKnownMavenArtifactId, allowing the separation of the two distinct
concepts "this is an artifact that Google provides for download"
and "this is an artifact that Android Studio knows something about.

Move the Kotlin artifacts out of GoogleMavenArtifactId.  Also correct
or remove some artifacts that don't in fact exist.

Add a displayName property which can and should be used to implement
`toString()` in implementing classes.  (Language semantics prevent us
from implementing `toString()` directly in the interface, sadly)

(The set of entries in GoogleMavenArtifactId is used to generate a
list of possible dependencies to add with the "Add Library Dependency"
intention action.  For artifacts in gmaven, Studio is able to use
online methods, or if necessary an offline fallback, to find suitable
versions to offer.  For arbitrary other artifacts not in gmaven,
Studio will offer, if found, the latest version present in the user's
Gradle cache, which might not be suitable at all.)

Bug: 279149611
Test: existing
Change-Id: I6cca815045138f07f5e5dfe3ab8dcbbd8e290d83