Update Gradle AGP checks to allow picking latest stable
First, the AGP out-of-date check would deliberately not
jump to the latest available version if there's a more
recent "micro" version available. The idea here was that
it would make each jump easier to do (e.g. if you're on 7.1.2,
and 7.1.4 is available, offer that first -- only offer
7.2.0 once you're already on the last 7.1.x version). But
that's not always desirable, so make that configurable.

Second, there was code in there to make sure we don't offer
an AGP version that is more recent than the corresponding
Studio version. This was relying on comparing Studio and
AGP version numbers; that used to work, but no longer does
now that we've decoupled the versioning schemes (e.g.
AGP 8.x and Studio 2025.x -- we're comparing 8 and 2025.
This CL improves this situation; we now pick up the IDE's
corresponding known AGP version.

Bug: 145606749
Test: Tests updated
Change-Id: I976cffe780e13bcc3ab3e57ae1bfb5e16c46eeb1