Fix `AdbServerControllerImpl` state machine cancellations
The AdbServerControllerImpl state machine should not throw
`CancellationExceptions` when a transition is interrupted by a
subsequent transition. For example, if a user calls `start`, and before
it completes, `stop` is called, the caller of `start` should receive an
appropriate error, rather than a `CancellationException`. This prevents
misinterpreting the interruption as the calling coroutine being
cancelled.

Bug: 378111505
Test: added
Change-Id: I0a64482d6615f70404c4098a16d6a1d7028fad06