Make ExifInterfaceDetector work on Kotlin sources
Extend the set of UElements the detector applies to in order to be
able to catch Kotlin constructors.  The underlying Kotlin Psi (and
indeed the UElements) do not allow for isolation of the qualified
method name in a method call expression: only the unqualified name (as
an identifier), or the whole call.  Given that constraint, prefer to
annotate the whole call so that later processing can (potentially)
replace it, rather than just the unqualified name (which would require
subsequent processing to do more involved surgery on the Psi).

Add a constant that will be useful for implementing the replacement by
new ExifInterface classes in Kotlin.

Bug: none filed
Test: added
Change-Id: Id890b2dcb0a8f5e1775d6d080be3ebf4dc5c8dd3