Propagate meta annotation of annotation on property even with specific use-sites
Previously, we only do so for annotation without use-sites
while propagating annotations on property to accessors, assuming
the intention is to apply the annotation to all the associated
members such as field and accessors.
That is, ironically, meta annotation on annotation with specific
use-site has been handled differently (i.e., not propagated).

Bug: 406850340
Test: unit test added
Change-Id: I2f51d6b94123c95aab0fffa0ec865408669bc3b7