[Built-in Kotlin] Support Kotlin compiler Gradle plugins
Kotlin compiler Gradle plugins interact with the build through the
org.jetbrains.kotlin.gradle.plugin.KotlinCompilation interface.

Implementing the full API surface is a bit challenging, so this change
implements the minimum set of APIs required for most plugins to work.

If a plugin uses an unsupported API, we'll either ask them to migrate to
supported APIs or consider supporting that API.

Bug: 409528883
Test: BuiltInKotlinCompilerPluginTest
Change-Id: I013503334a6759720c6457c0c32ea7aac6eea66e