adblib: Refactor `TrackApp` and `TrackJdwp`
This change cleans up and makes code related to
the `TrackApp` and `TrackJdwp` classes more
consistent with each other.

* Both expose a `StateFlow` as a property
* Both StateFlow expose a `ListWithFStateFlowStatus` that provides
access to all elements in the list in addition to a "current state
flow status" property to provide extra information about the state
of the collector (e.g. "just started", "inactive", "retrying connection",
etc.)

Bug: n/a
Test: Existing
Change-Id: I34086e035ed4d63a7f15d07d586497dffe483ccd