Don't suggest higher preview versions with a lower channel stability
If you're currently using 1.4.0-beta01, and the versions 1.4.0 and
1.5.0-alpha01 are available, lint will only suggest updating to 1.4.0,
the latest stable. However, if there is no stable version released yet,
it would suggest updating to 1.5.0-alpha01. But we shouldn't offer
upgrades to a lower quality channel (e.g. alpha rather than beta). This
CL fixes that.

It also moves the update filter out of the detector instance and into
the companion object such that it can be used from other classes calling
into the version comparison methods in the companion object.

Bug: Not filed
Test: Unit test included
Change-Id: Ia161d37c82ee744f9f0e2e17235869d73a2df14c