[Built-in Kotlin] Use KGP's factory to create KotlinCompilation
In the previous commit (5663e8d), we provided a custom implementation of
the KotlinCompilation interface to support Kotlin compiler plugins.

A better approach would be to reuse KGP's existing implementation. Since
the API to do that is not yet available (tracked by KT-77023), for now
we will try to make use of KotlinJvmAndroidCompilationFactory via
reflection.

Bug: 409528883
Test: Updated BuiltInKotlinCompilerPluginTest
Change-Id: Icb9cef0f402a587e1a4fa17b5a41acc8bdf5368c