Add GradleVersion as part of ProjectTemplateData
The ProjectTemplateDataBuilder is now responsible given specified AGP
version to find a compatible Gradle version and add it to the model.

This removes responsibility from ProjectTemplateRenderer but also
ensures that ProjectTemplateData contains all project required info
for the recipe given that Gradle version will be required to configure
Gradle settings for Daemon toolchain by applying the toolchain resolver
plugin. At the same time Gradle version might be useful in the future
when defining other DSL configurations.

Bug: 388787307
Test: Existing
Change-Id: I2d07e3466c32502866c277ee7d48d823c3305078