Do NOT include Views from WindowInspector in xr views
This fixes a problem with applications using XR libraries before
1.0.0-alpha04.

We call XrViewsProvider.getXrViews() which with XR libraries
before 1.0.0-alpha04 is unable to retrieve the XR views.
Since we include the Views from "full space mode" we fail to
fallback to the method using reflection.
Force the fallback by NOT returning the "full space mode".

Fixes: 416905652
Test: N/A
Change-Id: I5d345f24f14869ed0fa2a6e476088ed37ebf33ad