Add new diff utilities
We've had needs for a "diff" utility in a number of places, usually in
tests, but not exclusively - and these have generally been duplicated to
avoid weird dependencies. This CL adds a central utility and (in related
CLs) replaces usages with it.

The older utilities also had various quirks that are fixed here; for
example, the diff hunks didn't generally have correct line numbers (or
correctly end with @@); another difference is that the old diff utils
generally put a separating space between the "+" or "-" and the line
content; git diff doesn't do that, and the new implementation doesn't
either. It also now includes line separators in the diff comparisons,
which matters for mixtures or \n and \r\n, as well as whether there
is a newline at the end of the file.

These CLs also updates the corresponding test output to match this (or
in the case of lint, where the diff output is part of the API, updates
diffs testing machinery to fall back on the old implementation.)

Bug: Not filed
Test: Unit tests included
Change-Id: I9eb6d9d07ca9462dc3ace39c9eaa007df25cf965