Update lint golden files
This CL updates the expected diff output for quickfix tests in the lint
unit tests to use the new canonical format.

(Lint falls back to retry the check with the old diff output, so this
change isn't required, but makes the tests do slightly less work. This
only updates the Kotlin files, which I could easily patch automatically;
it leaves the Java files alone, continuing to test the compatibility
handling.)

Bug: None
Test: This is it
Change-Id: I2029e8efb0f5dd81c0f0254af2ba9d52233d0298