created public interface to retrieve junit engine inputs.
The API is actually a class since we do not want to have junit
engines depending on the AGP implementation to just get this.
Alternatively, we could consider using a new module all together
to store public APIs for test engines but at this point, it seems
to be an overkill so we will reconsider later.

API is simple and allow having access to all requested parameters
without dealing with the environent variable nor the file format.
The file format used to be JSON but it is now simplified to just
be a serialized Properties. File format might change in the future
again.

Test: adapted existing test to use public APIs
Bug: N/A
RelNote: N/A
Change-Id: Ib289a314754ef4c65207e86d5cfed78e83cc1284