Fix findRuleJars
...to add lint jars from compiled dependencies.

This appears to have stopped working for global analysis that includes
all code (e.g. via Inspect code... in Android Studio) since:

Ic1a584a1e3a9955464cbc85533326776566dabca

Previously, every compiled dependency in the IDE was associated with a
Lint Project. Without this, findRuleJars will only include lint jars
from compiled dependencies if we are analyzing a subset of the project.
This change makes it so we always add the jars.

Bug: 424430560
Test: in I75e2ad563b66e35881fe6499a51cf353c395c2a5
Change-Id: Iab2d56157a8df3e996e233f200601f6eb4851813