426554247: TypoDetector regularly crashes with IllegalArgumentException during quickfix
This fixes the range computation for text nodes such that it properly
updates the ending location as well. Before this, it was only updating
the offset, not the column. This meant that various error location
output was only showing a single character as the error range (^)
instead of the whole range.

This was hiding the fact that in the typo detector was setting the wrong
ending offset for text nodes -- which meant that the associated range
for replacement quickfixes was missing the complete range, which meant
that quickfix replacement computation could fail finding the original
to-be-replaced string.

Bug: 426554247
Test: Tests included
Change-Id: I03dbf60d57e4e4e7248a474a230870d1051b39fe