[fused lib] Fix NPE when aar-metadata.properties does not specify minCompileSdk
aarMetdata merging causes a NullPointerException when an
aar-metadata.properties file exists in a dependency, but does not
specify minCompileSdk or minCompileSdkExtension as a property.

Change also slightly changes merging of minSdkCompileSdkExtension. Before it selects the highest from all dependencies, however, it will now select the highest from the highest minCompileSdk aarMetadata.

Fixes: 426156521
Test: FusedLibraryMergeArtifactsTest
Change-Id: I8842f826378bd79e6a0db586c5bd817d28a78bb1