Fix support for newline in `.bazelversion`
Bazel itself expects a newline character at the end of files, as does
Gerrit, but there was a bug in our Bazel wrapper where we didn't strip
newlines when reading the file. The Windows shell was less happy about
this and would fail to pass the correct arguments to `bazel`.

This CL ensures that we strip trailing whitespace and adds the newline.

See also:
https://github.com/bazelbuild/bazel/blob/7fb19f48054eba275d9cf09f31cb548457fc72e3/scripts/packages/bazel.sh#L118

Bug: 402848371
Test: studio-win target succeeds
Change-Id: I9e23f242b445fd092a231f747d351cea942b18d1